requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-admin-honeypot==1.1.0
django-admin-interface==0.14.1
django-colorfield==0.3.2
django-debug-toolbar==3.1.1
django-defender==0.8.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.4.0
django-lockdown==3.0.0
django-loginas==0.3.9
django-multi-captcha-admin==1.0.0
django-recaptcha2==1.4.1
django-request==1.5.6
django-username-email==2.4.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==3.3.0
jdcal==1.4.1
kombu==5.0.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
prompt-toolkit==3.0.9
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
django-filter